requirements.txt
argcomplete==1.10.0
beautifulsoup4==4.8.0
bs4==0.0.1
bz2file==0.98
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.6
django-cors-headers==3.1.1
django-filter==2.2.0
djangorestframework==3.10.3
docx2txt==0.8
EbookLib==0.17.1
epub==0.5.2
epub-conversion==1.0.7
et-xmlfile==1.0.1
extract-msg==0.23.1
html2text==2019.9.26
idna==2.8
IMAPClient==2.1.0
jdcal==1.4.1
lxml==4.4.1
nltk==3.4.5
nose==1.3.7
numpy==1.17.3
odfpy==1.4.0
olefile==0.46
openpyxl==3.0.0
pandas==0.25.2
pdfminer.six==20181108
Pillow==6.2.0
pycryptodome==3.9.0
PyPDF2==1.26.0
python-dateutil==2.8.0
python-docx==0.8.10
python-pptx==0.6.18
pytz==2019.3
Random-Word==1.0.4
requests==2.22.0
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.9.4
SpeechRecognition==3.8.1
sqlparse==0.3.0
striprtf==0.0.5
textract==1.6.3
tzlocal==1.5.1
urllib3==1.25.6
xlrd==1.2.0
XlsxWriter==1.2.2
xml-cleaner==2.0.4
